refactor!: move hydration store to separate module
