+++ src/core/init/helpers.ts
@@ -18,7 +18,8 @@ import HydrationStore from 'core/hydration-store';
 import * as net from 'core/net';
 import * as cookies from 'core/cookies';
 
-import { HydrationStore, State } from 'core/component';
+import type { State } from 'core/component';
+import { HydrationStore } from 'core/hydration-store';
 import type { InitAppOptions, CreateAppOptions } from 'core/init/interface';
 
 /**
